// libraries/common-ui/src/types.d.ts
export type TooltipType = "mouse" | "elcenter" | "elside" | "elupdown";
// ...
export type ChartValue = string | number | Date;
// ...
export type Aggregation =
  | "count"
  | "sum"
  | "min"
  | "max"
  | "avg"
  | "median"
  | "deviation"
  | "variance"
  | "quantileThree"
  | "quantileOne";

export type Location = "top" | "right" | "bottom" | "left" | "center";

export type Position = "start" | "middle" | "end";

// hanging = top, ideographic = bottom, middle = central
export type BaseLine = "hanging" | "ideographic" | "central";

export type Anchor = "start" | "middle" | "end";

export type CollisionTextType = "back" | "series";

export interface IChartData {
  [key: string]: any;
  key: ChartValue;
}

export interface INonNullMargin {
  [key: string]: any;
  top: number;
  right: number;
  bottom: number;
  left: number;
}

export interface IMargin {
  [key: string]: any;
  top?: number;
  right?: number;
  bottom?: number;
  left?: number;
}

export interface IChartValue {
  idx: number;
  keyData: ChartValue;
  valueData: number;
  valueField: string;
  outlier?: boolean;
  highlight?: boolean;
}

export interface IChartAggregation {
  valueData: number;
  valueField: string;
}

export interface IChartTooltipParams {
  type: "key" | "value";
  key: ChartValue;
  text: ChartValue;
  numberFormat?: (value: number, percent?: boolean) => string;
}